move package tests to dedicated package
